+++ data/src/main/java/com/juco/data/local/entity/WorkPlaceEntity.kt
@@ -7,7 +7,7 @@ import androidx.room.PrimaryKey
 data class WorkPlaceEntity(
     @PrimaryKey(autoGenerate = true) val id: Int = 0,
     val name: String,
-    val wage: Int,
+    val wage: Long,
     val workDays: String,
     val payDay: String
 )
